+++ src/utils/OverlayServiceModule.ts
@@ -3,6 +3,8 @@ import { NativeModules } from 'react-native';
 interface OverlayServiceModuleType {
   startOverlayService(): Promise<boolean>;
   stopOverlayService(): Promise<boolean>;
+  setButtonOpacity(opacity: number): Promise<boolean>;
+  getButtonOpacity(): Promise<number>;
 }
 
 const OverlayServiceModule: OverlayServiceModuleType = NativeModules.OverlayServiceModule;